Complex types - mixed list

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@ A module implementation of complex types-mixed list
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float | int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int | float]]) -> float:
     """
     Parameters
     ----------
